Move code into stats.py
